final fix

+++ src/main/java/ru/practicum/shareit/user/services/UserServiceImpl.java
@@ -1,13 +1,17 @@
 package ru.practicum.shareit.user.services;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 import ru.practicum.shareit.handler.exceptions.UserNotFound;
 import ru.practicum.shareit.user.dto.UserDto;
+import ru.practicum.shareit.user.mapper.UserMapper;
 import ru.practicum.shareit.user.model.User;
 import ru.practicum.shareit.user.repositoryes.UserStorage;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 //TODO Разобраться как правильно указывать конструкторы
 @Service
@@ -16,28 +20,78 @@ public class UserServiceImpl implements UserService {
 
     private final UserStorage userStorage;
 
+    private static long userID = 0;
+
     @Override
-    public User save(UserDto userDto) {
-        return userStorage.save(userDto);
+    public UserDto save(UserDto userDto) {
+        if (userDto.getEmail() != null && !userDto.getEmail().isEmpty()) {
+            User checkUser =
+                    userStorage.findAll().stream()
+                            .filter(user -> user.getEmail().equals(userDto.getEmail()))
+                            .findFirst()
+                            .orElse(null);
+            if (checkUser == null) {
+                User user = UserMapper.toUser(userDto);
+                user.setId(++userID);
+                return UserMapper.toUserDto(userStorage.save(user));
+            } else {
+                throw new RuntimeException("Duplicated email");
+            }
+        } else {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        }
     }
 
     @Override
-    public User put(UserDto userDto, long userId) throws UserNotFound {
-        return userStorage.put(userDto, userId);
+    public UserDto put(UserDto userDto, long userId) throws UserNotFound {
+        if (userDto != null && userStorage.findById(userId) != null) {
+            User checkUser =
+                    userStorage.findAll().stream()
+                            .filter(user -> user.getEmail().equals(userDto.getEmail()))
+                            .findFirst()
+                            .orElse(null);
+            if (checkUser == null) {
+                User user = userStorage.findById(userId);
+                user.setId(userId);
+                if (userDto.getName() != null) {
+                    user.setName(userDto.getName());
+                }
+                if (userDto.getEmail() != null) {
+
+                    user.setEmail(userDto.getEmail());
+                }
+                userStorage.put(user, userId);
+                return UserMapper.toUserDto(user);
+            } else {
+                throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR);
+            }
+        } else {
+            throw new UserNotFound(String.format("User %s not found", userId));
+        }
     }
 
     @Override
     public void delete(long userId) throws UserNotFound {
-        userStorage.delete(userId);
+        if (userStorage.findById(userId) != null) {
+            userStorage.delete(userId);
+        } else {
+            throw new UserNotFound(String.format("User %s not found", userId));
+        }
     }
 
     @Override
-    public List<User> findAll() {
-        return userStorage.findAll();
+    public List<UserDto> findAll() {
+        return userStorage.findAll().stream()
+                .map(user -> new UserDto(user.getId(), user.getName(), user.getEmail()))
+                .collect(Collectors.toList());
     }
 
     @Override
-    public User findById(long userId) throws UserNotFound {
-        return userStorage.findById(userId);
+    public UserDto findById(long userId) throws UserNotFound {
+        try {
+            return UserMapper.toUserDto(userStorage.findById(userId));
+        } catch (UserNotFound e) {
+            throw new UserNotFound(String.format("User %s not found", userId));
+        }
     }
 }
